third_party/WebKit/LayoutTests/fast/css/vertical-align-inherit.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
#parent1 { vertical-align: 50px; }
#parent2 { vertical-align: 50%; }
#child1, #child2 { vertical-align: inherit; }
</style>
<div id="parent1">
  <div id="child1"></div>
</div>
<div id="parent2">
  <div id="child2"></div>
</div>
<script>
test(function() {
    assert_equals(getComputedStyle(child1).verticalAlign, "50px");
    assert_equals(getComputedStyle(child2).verticalAlign, "50%");
}, "Tests vertical-align inherits correctly");
</script>
